Fix integer overflows in size_t calculations

+++ src/iop/velvia.c
@@ -134,7 +134,7 @@ void process(struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, const 
 
   // Apply velvia saturation
   if(strength <= 0.0)
-    memcpy(ovoid, ivoid, (size_t)sizeof(float) * ch * roi_out->width * roi_out->height);
+    memcpy(ovoid, ivoid, sizeof(float) * ch * roi_out->width * roi_out->height);
   else
   {
 #ifdef _OPENMP
@@ -182,7 +182,7 @@ void process_sse2(struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, c
 
   // Apply velvia saturation
   if(strength <= 0.0)
-    memcpy(out, in, (size_t)sizeof(float) * ch * roi_out->width * roi_out->height);
+    memcpy(out, in, sizeof(float) * ch * roi_out->width * roi_out->height);
   else
   {
 #ifdef _OPENMP
